+++ tests/console/commands/cache/test_clear.py
@@ -1,65 +1,17 @@
 from __future__ import annotations
 
-import uuid
-
 from typing import TYPE_CHECKING
 
 import pytest
-
-from cachy import CacheManager
 
 
 if TYPE_CHECKING:
     from pathlib import Path
 
-    from _pytest.monkeypatch import MonkeyPatch
+    from cachy import CacheManager
     from cleo.testers.command_tester import CommandTester
 
     from tests.types import CommandTesterFactory
-
-
-@pytest.fixture
-def repository_cache_dir(monkeypatch: MonkeyPatch, tmpdir: Path) -> Path:
-    from pathlib import Path
-
-    import poetry.locations
-
-    path = Path(str(tmpdir))
-    monkeypatch.setattr(poetry.locations, "REPOSITORY_CACHE_DIR", path)
-    return path
-
-
-@pytest.fixture
-def repository() -> str:
-    return f"01_{uuid.uuid4()}"
-
-
-@pytest.fixture
-def cache(repository_cache_dir: Path, repository: str) -> CacheManager:
-    cache = CacheManager(
-        {
-            "default": repository,
-            "serializer": "json",
-            "stores": {
-                repository: {
-                    "driver": "file",
-                    "path": str(repository_cache_dir / repository),
-                }
-            },
-        }
-    )
-    return cache
-
-
-@pytest.fixture
-def mock_caches(
-    repository_cache_dir: Path,
-    repository: str,
-    cache: CacheManager,
-) -> None:
-    (repository_cache_dir / repository).mkdir()
-    cache.remember_forever("cachy:0.1", lambda: {"name": "cachy", "version": "0.1"})
-    cache.remember_forever("cleo:0.2", lambda: {"name": "cleo", "version": "0.2"})
 
 
 @pytest.fixture
@@ -69,39 +21,40 @@ def tester(command_tester_factory: CommandTesterFactory):
 
 def test_cache_clear_all(
     tester: CommandTester,
-    repository: str,
+    repository_one: str,
     repository_cache_dir: Path,
-    mock_caches: None,
+    cache: CacheManager,
 ):
-    tester.execute(f"{repository} --all", inputs="yes")
+    tester.execute(f"{repository_one} --all", inputs="yes")
 
     assert tester.io.fetch_output() == ""
     # ensure directory is empty
-    assert not any((repository_cache_dir / repository).iterdir())
+    assert not any((repository_cache_dir / repository_one).iterdir())
+    assert not cache.has("cachy:0.1")
+    assert not cache.has("cleo:0.2")
 
 
 def test_cache_clear_all_no(
     tester: CommandTester,
-    repository: str,
+    repository_one: str,
     repository_cache_dir: Path,
     cache: CacheManager,
-    mock_caches: None,
 ):
-    tester.execute(f"{repository} --all", inputs="no")
+    tester.execute(f"{repository_one} --all", inputs="no")
 
     assert tester.io.fetch_output() == ""
     # ensure directory is not empty
-    assert any((repository_cache_dir / repository).iterdir())
+    assert any((repository_cache_dir / repository_one).iterdir())
+    assert cache.has("cachy:0.1")
+    assert cache.has("cleo:0.2")
 
 
 def test_cache_clear_pkg(
     tester: CommandTester,
-    repository: str,
-    repository_cache_dir: Path,
+    repository_one: str,
     cache: CacheManager,
-    mock_caches: None,
 ):
-    tester.execute(f"{repository}:cachy:0.1", inputs="yes")
+    tester.execute(f"{repository_one}:cachy:0.1", inputs="yes")
 
     assert tester.io.fetch_output() == ""
     assert not cache.has("cachy:0.1")
@@ -110,12 +63,11 @@ def test_cache_clear_pkg(
 
 def test_cache_clear_pkg_no(
     tester: CommandTester,
-    repository: str,
-    repository_cache_dir: Path,
+    repository_one: str,
     cache: CacheManager,
-    mock_caches: None,
 ):
-    tester.execute(f"{repository}:cachy:0.1", inputs="no")
+    tester.execute(f"{repository_one}:cachy:0.1", inputs="no")
 
     assert tester.io.fetch_output() == ""
     assert cache.has("cachy:0.1")
+    assert cache.has("cleo:0.2")
